NCI-Agency/anet#379: Expands simulator functionality
- Repaired buildup mechanism which inserts a determinate amount of data into the database.
- Expanded the buildup mechanism to also take optional arguments into account.
- Kept the stories mechanism as an option.
- Reworked the scenarios to generalize the setup and easily add new ones.
- Added logging for improved feedback.
- Reworked old scenarios into ‘scenario1’ and ‘scenario2’.
- Added ‘scenario3’ which generates a large amount of data.
