+++ client/tests/sim/scenarios/scenario1.js
@@ -1,0 +1,12 @@
+import userTypes from "./data/users/userTypes"
+import buildup from "./data/buildup/oldScenario"
+import stories from "./data/stories/storySim"
+
+// Original scenario which was not in use anymore
+const scenario1 = {
+  userTypes: userTypes,
+  buildup: buildup,
+  stories: stories
+}
+
+export default scenario1
