+++ client/tests/sim/scenarios/scenario2.js
@@ -1,0 +1,11 @@
+import userTypes from "./data/users/userTypes"
+import stories from "./data/stories/storySim"
+
+// Original scenario which runs stories for ma certain ammount of time
+const scenario2 = {
+  userTypes: userTypes,
+  buildup: [],
+  stories: stories
+}
+
+export default scenario2
