+++ client/tests/sim/scenarios/scenario3.js
@@ -1,0 +1,11 @@
+import userTypes from "./data/users/userTypes"
+import buildup from "./data/buildup/lotsOfData"
+
+// New scenario intended to load the database with a large amount of data
+const scenario3 = {
+  userTypes: userTypes,
+  buildup: buildup,
+  stories: []
+}
+
+export default scenario3
